+++ nodalia-cards.manifest.js
@@ -1,10 +1,10 @@
 export default {
-  "pkgVersion": "1.1.0-alpha.20",
-  "contentSha256_12": "d2f76dd9aca6",
+  "pkgVersion": "1.1.0-alpha.21",
+  "contentSha256_12": "0988e52db662",
   "file": "nodalia-cards.bundle.js",
   "loaderFile": "nodalia-cards.js",
-  "hacsFile": "nodalia-cards-1.1.0-alpha.20.js"
+  "hacsFile": "nodalia-cards-1.1.0-alpha.21.js"
 };
-export const pkgVersion = "1.1.0-alpha.20";
-export const contentSha256_12 = "d2f76dd9aca6";
+export const pkgVersion = "1.1.0-alpha.21";
+export const contentSha256_12 = "0988e52db662";
 export const file = "nodalia-cards.bundle.js";